pyproject.toml
[project]
name = "fastapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.14",
    "locust>=2.37.11",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pydantic-settings>=2.10.1",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.35.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
